requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
python-dotenv
pandas>=2.2.0
PyPDF2
chromadb
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-openai
transformers
torch
sentence-transformers
streamlit
requests
pydantic
typing-extensions
